backend/requirements.txt
asgiref==3.11.1
Django==6.0.4
django-cors-headers==4.9.0
django-environ==0.13.0
django-stubs==6.0.3
django-stubs-ext==6.0.3
djangorestframework==3.17.1
djangorestframework-stubs==3.16.9
factory_boy==3.3.3
Faker==40.15.0
psycopg2-binary==2.9.11
redis==7.4.0
sqlparse==0.5.5
types-PyYAML==6.0.12.20260408
typing_extensions==4.15.0
whitenoise==6.12.0
